+++ CHANGELOG.md
@@ -32,6 +32,18 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 * old `PFixedDecimal` is renamed and relocated into `Plutarch.Extra.Fixed`.
 
+## 3.11.1 -- 2022-10-27
+
+### Added
+
+* Modifier `GenAssetClass` to provide QuickCheck support for `AssetClass`
+* Helper type `AdaClassPresence` for indicating whether `GenAssetClass` should
+  generate the ADA class or not
+
+### Modified
+
+* `plutarch-quickcheck` is now a direct dependency, rather than test only.
+
 ## 3.11.0 -- 2022-10-25
 
 ### Added
@@ -46,11 +58,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## 3.10.4 -- 2022-10-25
 
-Changes to `Plutarch.Extra.AssetClass`:
+### Changed
 
-* Remove unnecessary `PAsData` wrappers
-* Allow tags of `AssetClass` to be poly-kinded
-* `PlyArg` instance for `AssetClass`
+* `Plutarch.Extra.AssetClass`:
+  * Remove unnecessary `PAsData` wrappers
+  * Allow tags of `AssetClass` to be poly-kinded
+  * `PlyArg` instance for `AssetClass`
 
 ## 3.10.3 -- 2022-10-24